refactor(pdf): extract shared types from pdfExport utilities into pdfExport.types.ts
